subcore: add prediction algorithm
subcore now guesses the user's load based on
their previous history. for example, if the
user's load requires load level 1 for 5 times in
a row, then add 3000ms to the scan time. this way,
subcore doesn't have to scan as often, which means
better battery, and better consistency for the user.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>